+++ cmd/containerd/builtins_devmapper_linux.go
@@ -18,4 +18,4 @@
 
 package main
 
-import _ "github.com/containerd/containerd/snapshots/devmapper"
+import _ "github.com/containerd/containerd/snapshots/devmapper/plugin"
